+++ apps/entropretty-gallery/src/components/Drawing.tsx
@@ -1,6 +1,7 @@
-import { EntroprettyEditorWorker, SchemaMetadata } from "entropretty-editor"
+import { EntroprettyEditorWorker } from "entropretty-editor"
+import type { SchemaMetadata } from "entropretty-utils"
 import { Remote } from "comlink"
-import { useEffect, useRef } from "react"
+import { useEffect, useMemo, useRef } from "react"
 
 interface Props {
   seed: Uint8Array
@@ -11,23 +12,23 @@ interface Props {
 
 export const Drawing: React.FC<Props> = ({ seed, schema, size, worker }) => {
   const canvasRef = useRef<HTMLCanvasElement>(null)
+  const canvasSize = useMemo(() => size * 2, [size])
 
   useEffect(() => {
     if (canvasRef.current === null) return
 
-    worker.drawTransfer(schema.name, seed, size).then((bitmap) => {
+    worker.drawTransfer(schema.name, seed, canvasSize).then((bitmap) => {
       const context = canvasRef.current!.getContext("2d")!
-      context.clearRect(0, 0, size, size)
-      context.drawImage(bitmap, 0, 0, size, size)
+      context.clearRect(0, 0, canvasSize, canvasSize)
+      context.drawImage(bitmap, 0, 0, canvasSize, canvasSize)
     })
   }, [seed, schema, size, worker])
 
   return (
     <canvas
-      title={schema.artist}
       ref={canvasRef}
-      width={size}
-      height={size}
+      width={canvasSize}
+      height={canvasSize}
       style={{ width: size, height: size }}
     />
   )