+++ packages/entropretty-editor/src/components/DrawingBitmap.tsx
@@ -1,45 +1,47 @@
-import { Skeleton } from "@/components/ui/skeleton"
 import { cn } from "@/lib/utils"
-import { useEffect, useRef, useState } from "react"
+import { useEffect, useMemo, useRef, useState } from "react"
 import { useApp } from "../lib/state"
 
 interface Props {
+  id?: string
   seed: Uint8Array
   schema: string
   size: number
+  scale?: number
 }
 
-export const DrawingBitmap: React.FC<Props> = ({ seed, schema, size }) => {
+export const DrawingBitmap: React.FC<Props> = ({
+  seed,
+  schema,
+  size,
+  scale = 2,
+  id,
+}) => {
   const [ready, setIsReady] = useState(false)
   const canvasRef = useRef<HTMLCanvasElement>(null)
   const worker = useApp((state) => state.worker)
 
+  const drawingSize = useMemo(() => size * scale, [size, scale])
+
   useEffect(() => {
     if (canvasRef.current === null) return
 
-    worker!.drawTransfer(schema, seed, size).then((bitmap) => {
+    worker!.drawTransfer(schema, seed, drawingSize).then((bitmap) => {
       const context = canvasRef.current!.getContext("2d")!
-      context.clearRect(0, 0, size, size)
-      context.drawImage(bitmap, 0, 0, size, size)
+      context.clearRect(0, 0, drawingSize, drawingSize)
+      context.drawImage(bitmap, 0, 0, drawingSize, drawingSize)
       setIsReady(true)
     })
-  }, [seed, schema, size, worker])
+  }, [seed, schema, size, worker, drawingSize])
 
   return (
-    <>
-      <Skeleton
-        title={seed.join(",")}
-        className={cn({ hidden: ready })}
-        style={{ width: size, height: size }}
-      />
-      <canvas
-        title={seed.join(",")}
-        className={cn("cursor-pointer", { hidden: !ready })}
-        ref={canvasRef}
-        width={size}
-        height={size}
-        style={{ width: size, height: size }}
-      />
-    </>
+    <canvas
+      id={id}
+      className={cn("cursor-pointer", { hidden: !ready })}
+      ref={canvasRef}
+      width={drawingSize}
+      height={drawingSize}
+      style={{ width: size, height: size }}
+    />
   )
 }