+++ packages/entropretty-editor/src/index.ts
@@ -1,4 +1,3 @@
-export type * from "./types"
 export { EntroprettyEditor } from "./EntroprettyEditor"
 export { createWorker } from "./lib/createWorker"
 export type { EntroprettyEditorWorker } from "./lib/createWorker"