+++ packages/entropretty-editor/src/lib/utils.ts
@@ -1,6 +1,6 @@
 import { clsx, type ClassValue } from "clsx"
 import { twMerge } from "tailwind-merge"
-import { FamilyKind } from "../types"
+import { FamilyKind } from "entropretty-utils"
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))