+++ packages/entropretty-utils/src/helpers.ts
@@ -10,7 +10,7 @@ export function contextPrelude(ctx: CanvasRenderingContext2D) {
 }
 
 export function split(seed: Uint8Array, parts: number) {
-  let r = []
+  let r: Array<number> = []
   let last = 0
   for (let i = 0; i < parts; ++i) {
     let next = Math.round(((i + 1) * 32) / parts)
@@ -30,11 +30,11 @@ export function bytesToNibbles(bytes: Uint8Array) {
   return nibbles
 }
 
-export function bit(seed: Uint8Array, i: number) {
-  return (seed[Math.floor(i / 4) % 8] >> i % 4) & 1
+export function bit(seed: Uint8Array, i: number): number {
+  return (seed[Math.floor(i / 8) % seed.length] >> i % 8) & 1
 }
 
-export function bits(seed: Uint8Array, from = 0, to = 32) {
+export function bits(seed: Uint8Array, from = 0, to = 32): number {
   let r = 0
   for (let i = from; i < to; ++i) {
     r = ((r << 1) | bit(seed, i)) >>> 0
@@ -63,9 +63,21 @@ export function strokeEach<E, T extends Array<E>>(
   ctx: CanvasRenderingContext2D,
 ) {
   array.forEach((element, index) => {
-    ctx.save()
+    ctx.beginPath()
     fn(element, index)
-    ctx.restore()
+    ctx.stroke()
+  })
+}
+
+export function fillEach<E, T extends Array<E>>(
+  array: T,
+  fn: (element: E, index: number) => void,
+  ctx: CanvasRenderingContext2D,
+) {
+  array.forEach((element, index) => {
+    ctx.beginPath()
+    fn(element, index)
+    ctx.fill()
   })
 }