+++ packages/entropretty-cli/CHANGELOG.md
@@ -1,5 +1,30 @@
 # entropretty-cli
 
+## 1.0.21
+
+### Patch Changes
+
+- ea3179a: update compiled code
+
+## 1.0.20
+
+### Patch Changes
+
+- d00d88b: fixes and features
+
+## 1.0.19
+
+### Patch Changes
+
+- makes typescript an explicit requirement
+
+## 1.0.18
+
+### Patch Changes
+
+- bd7ce8b: release preperations
+- fd65182: adds `build` command
+
 ## 1.0.17
 
 ### Patch Changes